Relands "Fix routing and layout issue" (#931)
* Revert "Revert "Fix routing and layout issue (#923)" (#930)"

This reverts commit 1e2dfea173299302c60b482d9fbd979caaaad6ed.

* Fix merge conflicts
